Separate crawler and browser
