+++ examples/emulate-device.js
@@ -1,6 +1,6 @@
 const HCCrawler = require('../lib/hccrawler');
 
-const hccrawler = new HCCrawler({
+HCCrawler.launch({
   evaluatePage: (() => ({
     title: $('title').text(),
     h1: $('h1').text(),
@@ -9,15 +9,11 @@ const hccrawler = new HCCrawler({
   onSuccess: (result => {
     console.log('onSuccess', result);
   }),
-});
-
-hccrawler.launch()
-  .then(() => {
+})
+  .then(hccrawler => {
     hccrawler.queue({ url: 'https://example.com', device: 'iPhone 6 Plus' });
     hccrawler.queue({ url: 'https://example.com', device: 'iPad' });
     hccrawler.queue({ url: 'https://example.com', device: 'Nexus 7' });
-    return hccrawler.onIdle();
-  })
-  .then(() => {
-    hccrawler.close();
+    hccrawler.onIdle()
+      .then(() => hccrawler.close());
   });
