+++ examples/skip-duplicates.js
@@ -1,7 +1,8 @@
 const HCCrawler = require('../lib/hccrawler');
 
 const requestedObj = {};
-const hccrawler = new HCCrawler({
+
+HCCrawler.launch({
   concurrency: 1,
   evaluatePage: (() => ({
     title: $('title').text(),
@@ -16,16 +17,12 @@ const hccrawler = new HCCrawler({
     if (requestedObj[options.url]) return false;
     return true;
   }),
-});
-
-hccrawler.launch()
-  .then(() => {
+})
+  .then(hccrawler => {
     hccrawler.queue('https://example.com');
     hccrawler.queue('https://example.net');
     hccrawler.queue('https://example.org');
     hccrawler.queue('https://example.com'); // The queue won't be requested
-    return hccrawler.onIdle();
-  })
-  .then(() => {
-    hccrawler.close();
+    hccrawler.onIdle()
+      .then(() => hccrawler.close());
   });
